Restructure PageboyViewController extensions
